requirements.txt
uvicorn==0.24.0.post1
transformers==4.36.0
onnxruntime-gpu==1.16.3
fastapi==0.105.0
onnx==1.15.0